Basic Story Generation
Using OpenAI to provide options to the user. The user is given 4 options as chakra UI buttons that add to the story. The options are then regenerated to continue.

Moving forward, I will need to restrict the amount of story data passed to the AI to save on costs and potentially help with creativity and randomness.